+++ packages/amplify-graphql-api-construct-tests/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.13.3](https://github.com/aws-amplify/amplify-category-api/compare/amplify-graphql-api-construct-tests@1.13.2...amplify-graphql-api-construct-tests@1.13.3) (2024-11-20)
+
+**Note:** Version bump only for package amplify-graphql-api-construct-tests
+
+## [1.13.2](https://github.com/aws-amplify/amplify-category-api/compare/amplify-graphql-api-construct-tests@1.13.1...amplify-graphql-api-construct-tests@1.13.2) (2024-11-19)
+
+**Note:** Version bump only for package amplify-graphql-api-construct-tests
+
 ## [1.13.1](https://github.com/aws-amplify/amplify-category-api/compare/amplify-graphql-api-construct-tests@1.13.0...amplify-graphql-api-construct-tests@1.13.1) (2024-11-14)
 
 ### Bug Fixes